% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labne_hm.R
\name{check_lhm_parameters}
\alias{check_lhm_parameters}
\title{Check LaBNE+HM parameters}
\usage{
check_lhm_parameters(net, gma, Temp, k.speedup, m.in, L.in, w)
}
\arguments{
\item{net}{The complex network to be embedded into hyperbolic space.}

\item{gma}{The network's scaling exponent.}

\item{Temp}{The network's temperature.}

\item{k.speedup}{A speedup heuristic will be applied to nodes of degree < k.speedup.}

\item{m.in}{The expected initial node degree.}

\item{L.in}{The internal link formation rate.}

\item{w}{The window considered by HyperMap to refine angles found by LaBNE.}
}
\value{
A list with all the revised parameters.
}
\description{
Function that makes sure that all parameters passed to labne_hm are valid.
}
