% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_model.R
\name{check_ps_parameters}
\alias{check_ps_parameters}
\title{Check PS model parameters}
\usage{
check_ps_parameters(N, avg.k, gma, Temp)
}
\arguments{
\item{N}{The number network nodes.}

\item{avg.k}{The target average node degree.}

\item{gma}{The network's scaling exponent.}

\item{Temp}{The network temperature.}
}
\value{
None
}
\description{
Function that makes sure that all parameters passed to ps.model are valid.
}
