% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{compare_clustering_to_er}
\alias{compare_clustering_to_er}
\title{Check if a given network is strongly clustered}
\usage{
compare_clustering_to_er(network, epochs = 100, label = "Given network")
}
\arguments{
\item{network}{igraph; The network of interest.}

\item{epochs}{integer; The number of ER graphs to consider in the comparison.}

\item{label}{character; A label for the given network in the resulting plot.}
}
\value{
A ggplot object comparing the clustering coefficient of the given network with the average of \code{epochs} ER graphs.
}
\description{
Compares the clustering of a given network with that of \code{epochs} ER graphs with the same number of nodes and edges.
}
\examples{
# Generate an artificial network with the PS model
net <- ps_model(500, 6, 2.5, 0)
# Compare its clustering coefficient with ER networks 
# with the same number of nodes and edges
compare_clustering_to_er(net$network, 100, "PS network")

}
\author{
Gregorio Alanis-Lobato \email{galanisl@uni-mainz.de}
}
