% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_conn_probs}
\alias{get_conn_probs}
\title{Check whether short hyperbolic distances are indicative of links}
\usage{
get_conn_probs(net, polar, bins = 10)
}
\arguments{
\item{net}{igraph; A complex network.}

\item{polar}{data frame; Polar coordinates (r, theta) of all nodes in the network.}

\item{bins}{numeric; The number of distance bins to consider (default = 10).}
}
\value{
A data frame with the two following elements:
\item{dist}{The considered distance bins.}
\item{prob}{the connection probabilities within each bin.}
}
\description{
Given a complex network and the hyperbolic coordinates of its nodes, compute the fraction of existing edges between all nodes whose pairwise
distances are within the considered bin.
}
\examples{
# Generate an artificial network with the PS model
net <- ps_model(500, 6, 2.5, 0)

# Get the connection probability curve and plot it
conn <- get_conn_probs(net$network, net$polar, 15)

plot(conn$dist, conn$prob, pch = 16, xlab = "Hyperbolic distance", ylab = "Connection probability")

}
\references{
Papadopoulos, F. et al. (2012) Popularity versus similarity in growing networks. \emph{Nature} 489(7417):537-40.
}
\author{
Gregorio Alanis-Lobato \email{galanisl@uni-mainz.de}
}
