% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_theoretical_conn_probs}
\alias{get_theoretical_conn_probs}
\title{Connection probabilities in the PS model}
\usage{
get_theoretical_conn_probs(bins, N, avg.k, gma, Temp)
}
\arguments{
\item{bins}{vector; The hyperbolic distance bins or steps in which theoretical connection probabilities are to be computed.}

\item{N}{integer; Number of network nodes.}

\item{avg.k}{numeric; Network's average node degree.}

\item{gma}{numeric; The network's scaling exponent gamma.}

\item{Temp}{numeric; The network temperature.}
}
\value{
A data frame with the two following elements:
\item{dist}{The considered distance bins.}
\item{prob}{the connection probabilities within each bin.}
}
\description{
Given a a set of network properties, compute connection probabilities according to 
the Popularity-Similarity model [Papadopoulos et al. 2012, Nature 489(7417):537-40].
}
\examples{
# Generate an artificial network with the PS model
net <- ps_model(500, 6, 2.5, 0.1)

# Get the real and theoretical connection probability curves and plot them
conn <- get_conn_probs(net$network, net$polar, 15)
theo <- get_theoretical_conn_probs(conn$dist, vcount(net$network), 6, 2.5, 0.1)

plot(conn$dist, conn$prob, pch = 16, 
     xlab = "Hyperbolic distance", ylab = "Connection probability")
points(theo$dist, theo$prob, pch = 16, 
     xlab = "Hyperbolic distance", ylab = "Connection probability", col = "red")
legend("topright", c("Real", "Theory"), pch = c(16, 16), col = c("black", "red"))

}
\references{
Papadopoulos, F. et al. (2012) Popularity versus similarity in growing networks. \emph{Nature} 489(7417):537-40.
}
\author{
Gregorio Alanis-Lobato \email{galanisl@uni-mainz.de}
}
