% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{greedy_route_packets}
\alias{greedy_route_packets}
\title{Analyse the navigability of a complex network}
\usage{
greedy_route_packets(net, polar, source, target, faulty = c())
}
\arguments{
\item{net}{igraph; A complex network with N nodes.}

\item{polar}{data frame; Polar coordinates (r, theta) of all nodes in the network.}

\item{source}{vector; A vector with one or more node indices representing sources.}

\item{target}{vector; A vector, the same size as \code{source}, with one or more node indices representing targets.}

\item{faulty}{vector; A vector with one or more node indices representing faulty system components (default = c()).}
}
\value{
A vector, the same size as \code{source} or \code{target}, with the hop stretch required to deliver the packet 
from each source to each target, or 0 if the packet was dropped.
}
\description{
Given a network and source and destination nodes, greedy route packets in hyperbolic space and record successful deliveries. If the packet
visits any of the nodes in \code{faulty}, it is dropped and the delivery flagged as unsuccessful (unless the faulty node is a target).
}
\examples{
# Generate an artificial network with the PS model
N <- 500
net <- ps_model(N, 6, 2.5, 0)

# Form vectors of random non-redundant source-target pairs
st <- 1000
# We subtract 1, because the formulae to go from linear upper 
# diagonal indexing to (i,j) are zero-based
k <- sample(N*(N-1)/2, st) - 1
sources <- (N - 2 - floor(sqrt(-8*k + 4*N*(N-1)-7)/2.0 - 0.5))
targets <- (k + sources + 1 - N*(N-1)/2 + (N-sources)*((N-sources)-1)/2)

# Back to 1-based indexing
sources <- sources + 1
targets <- targets + 1

# Analyse the network's navigability
hop.stretch <- greedy_route_packets(net$network, net$polar, sources, targets)

# Compute the fraction of succesfully delivered packets
sum(hop.stretch)/st

}
\references{
Boguna, M. and Krioukov, D. (2009) Navigating Ultrasmall Worlds in Ultrashort Time. \emph{Physical Review Letters} 102(5).

Krioukov, D. et al. (2010) Hyperbolic geometry of complex networks. \emph{Physical Review E} 82(3).
}
\author{
Gregorio Alanis-Lobato \email{galanisl@uni-mainz.de}
}
