% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{hyperbolic_dist}
\alias{hyperbolic_dist}
\title{Hyperbolic distance between points}
\usage{
hyperbolic_dist(ZI, ZJ)
}
\arguments{
\item{ZI}{single-element data frame; A node in hyperbolic space placed at polar coordinates (\code{ZI$r}, \code{ZI$theta}).}

\item{ZJ}{data frame; m entries representing m nodes in hyperbolic space placed at polar coordinates (\code{ZJ$r}, \code{ZJ$theta}).}
}
\value{
An \code{m}-element vector with the hyperbolic distance between node \code{ZI} and nodes \code{ZJ}.
}
\description{
Computes the hyperbolic distance between a point at polar coordinates (\code{ZI$r}, \code{ZI$theta}) and \code{m} points at polar coordinates (\code{ZJ$r}, \code{ZJ$theta}).
}
\examples{
# Generate an artificial network with the PS model
net <- ps_model(500, 6, 2.5, 0)
# If nodes data frame contains m nodes, use the following 
# to compute the distance between node 1 and the rest (including itself):
nodes <- net$polar
d <- hyperbolic_dist(nodes[1,], nodes)
# To compute the pairwise distances between m nodes use:
D <- sapply(seq(nrow(nodes)), function(i) hyperbolic_dist(nodes[i, ], nodes))

}
\references{
Krioukov, D. et al. (2010) Hyperbolic geometry of complex networks. \emph{Physical Review E} 82(3).
}
\author{
Gregorio Alanis-Lobato \email{galanisl@uni-mainz.de}
}
