% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labne_hm.R
\name{labne_hm}
\alias{labne_hm}
\title{Laplacian-Based Network Embedding+HyperMap (LaBNE+HM)}
\usage{
labne_hm(net, gma = NA, Temp = 0.1, k.speedup = 10, m.in = NA,
  L.in = NA, w = "auto")
}
\arguments{
\item{net}{igraph, data frame or path to tab-separated file; The complex network to be embedded into hyperbolic space.}

\item{gma}{numeric; The network's scaling exponent. If not specified, it is automatically computed.}

\item{Temp}{numeric; The network's temperature (low temperature for strongly clustered networks and vice versa). If not specified, it is set to 0.1.}

\item{k.speedup}{integer; A speedup heuristic will be applied to nodes of degree < k.speedup. Its default value is 10. If set to 0, no heuristic is applied.}

\item{m.in}{numeric; The expected initial node degree, i.e. the average number of link stubs with which a node joins the network.  Parameter m.in can be obtained from historical data of the evolution of the network. If this data is not available and m.in is not specified, it is set to the minimum observed node degree in the network.}

\item{L.in}{numeric; The internal link formation rate, i.e. the average number of links formed between existing network nodes. If not specified, it is set to L = (kbar-2*m)/2, where kbar is the average node degree of the network.}

\item{w}{numeric; The window considered by HyperMap to refine angles found by LaBNE. If not specified, it is set to "auto", which means that w is set to 2*pi*Temp^2.}
}
\value{
List with the two following elements:
\item{network}{igraph object representation of the input network.}
\item{polar}{Data frame containing elements \code{r} and \code{theta}, the inferred radial and angular coordinates of the network nodes in hyperbolic space.}
\item{cartesian}{Data frame containing elements \code{x} and \code{y}, the inferred cartesian coordinates of the network nodes in hyperbolic space. It is useful to have these coordinates for the direct visualisation of the embedding using standard plotting function in R.}
}
\description{
LaBNE+HM is a method for the fast yet accurate embedding of complex networks to the native representation of the two-dimensional hyperbolic plane.
}
\examples{
# Generate an artificial network with the PS model, 
# such that the hyperbolic coordinates of its nodes are known
net <- ps_model(N = 500, avg.k = 10, gma = 2.5, Temp = 0.15)

# Map the network to hyperbolic space using LaBNE+HM and specifying all parameters.
coords <- labne_hm(net = net$network, gma = 2.5, Temp = 0.15, 
                   k.speedup = 10, m.in = 5, L.in = 0, w = pi/12)

# Visually explore the resulting hyperbolic mapping
plot_hyperbolic_net(network = net$network, 
                    nodes = coords$polar, node.colour = net$polar$theta)

}
\references{
Alanis-Lobato, G., Mier, P. and Andrade-Navarro, M. (2016) Manifold learning and maximum likelihood estimation for hyperbolic network embedding. \emph{Applied Network Science} 1(10).

Alanis-Lobato, G., Mier, P. and Andrade-Navarro, M. (2016) Efficient embedding of complex networks to hyperbolic space via their Laplacian. \emph{Scientific Reports} 6, 30108.
}
\author{
Gregorio Alanis-Lobato \email{galanisl@uni-mainz.de}
}
