% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{log_seq}
\alias{log_seq}
\title{Generate a log10-linearly spaced sequence}
\usage{
log_seq(s, e, n)
}
\arguments{
\item{s}{numeric; The sequence starting point.}

\item{e}{numeric; The sequence ending point.}

\item{n}{integer; The number of points to be generated.}
}
\value{
A numeric vector with \code{n} logarithmically spaced points between \code{s} and \code{e}.
}
\description{
Generates a sequence of \code{n} logarithmically spaced points between \code{s} and \code{e}, inclusive.
}
