% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{plot_hyperbolic_net}
\alias{plot_hyperbolic_net}
\title{Plot a network in hyperbolic space}
\usage{
plot_hyperbolic_net(network, nodes, node.colour = 1)
}
\arguments{
\item{network}{igraph; The network to be plotted.}

\item{nodes}{data frame; Polar coordinates (r, theta) of all nodes in the network.}

\item{node.colour}{vector; An optional parameter that allows to colour nodes according to a colour vector or a single colour.}
}
\value{
A ggplot of the given network in hyperbolic space.
}
\description{
Plots a network in hyperbolic space, given node polar coordinates and colours.
}
\examples{
# Generate an artificial network with the PS model
net <- ps_model(500, 6, 2.5, 0)
# Plot the network and colour nodes according to their radial coordinates
plot_hyperbolic_net(net$network, net$polar, net$polar$theta)

}
\author{
Gregorio Alanis-Lobato \email{galanisl@uni-mainz.de}
}
