% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{plot_hyperbolic_nodes}
\alias{plot_hyperbolic_nodes}
\title{Plot the nodes of a network in hyperbolic space}
\usage{
plot_hyperbolic_nodes(nodes, node.colour)
}
\arguments{
\item{nodes}{data frame;  Polar coordinates (r, theta) of all nodes in the network.}

\item{node.colour}{vector; An optional parameter that allows to colour nodes according to a colour vector or a single colour.}
}
\value{
A ggplot of the given network nodes in hyperbolic space.
}
\description{
Plots nodes of a network in hyperbolic space, given node polar coordinates and colours.
}
\examples{
# Generate an artificial network with the PS model
net <- ps_model(500, 6, 2.5, 0)
# Plot the network nodes and colour them according to their radial coordinates
plot_hyperbolic_nodes(net$polar, net$polar$r)

}
\author{
Gregorio Alanis-Lobato \email{galanisl@uni-mainz.de}
}
