% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_model.R
\name{ps_model}
\alias{ps_model}
\title{Generate a Popularity-Similarity network}
\usage{
ps_model(N = 500, avg.k = 10, gma = 2, Temp = 0)
}
\arguments{
\item{N}{integer; The number of desired nodes for the network (default = 500).}

\item{avg.k}{numeric; The target average node degree (default = 10).}

\item{gma}{numeric; The target scaling exponent of the network's node degree distribution (default = 2).}

\item{Temp}{numeric; The network temperature, controlling the target clustering coefficient (default = 0).}
}
\value{
List with the two following elements:
\item{network}{igraph object representation of the input network.}
\item{polar}{Data frame containing elements \code{r} and \code{theta}, the radial and angular coordinates of the network nodes in hyperbolic space.}
}
\description{
Generates an artificial network following the Popularity-Similarity model proposed in [Papadopoulos et al. 2012, Nature 489(7417):537-40].
This implementation only considers the hyperbolic space curvature \eqn{K = -(1^2)}.
}
\examples{
net.default <- ps_model()
net <- ps_model(100, 6, 2.5, 0.2)

}
\references{
Papadopoulos, F. et al. (2012) Popularity versus similarity in growing networks. \emph{Nature} 489(7417):537-40.
}
\author{
Gregorio Alanis-Lobato \email{galanisl@uni-mainz.de}
}
